Add files via upload
Floyd-Warshall is a shortest-path algorithm. It works with both directed and undirected graphs, it does not work with graphs with negative cycles though. 
Input: an adjacency matrix
Output: the shortest distance between each pair of vertices.

Time Complexity
There are three loops for computing the shortest path in the graph and each of these loops has constant complexities. Therefore, due to this, the time complexity of the Floyd Warshall algorithm is O(n3). Also, the space complexity of the Floyd Warshall algorithm is O(n2).